refactor(cart): change data type conversion quantity and update logic check stork is valid when add cart

+++ src/main/java/com/nguyenhuugiap/computer_shop/dto/cart/CartItemCreationRequest.java
@@ -16,5 +16,5 @@ public class CartItemCreationRequest {
 
     @NotNull(message = "QUANTITY_REQUIRED")
     @Min(value = 0, message = "QUANTITY_INVALID")
-    Integer quantity;
+    Long quantity;
 }
